requirements.txt
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-PyMongo==2.3.0
itsdangerous==1.1.0
pymongo==3.9.0
python-dotenv==0.10.3
Werkzeug==0.16.0
WTForms-2.2.1
flask-wtf-0.14.2
